fix: suspense queries
